Get rid of KC edit config page
