+++ kafka-ui-react-app/src/components/Connect/Connect.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { Navigate, Routes, Route } from 'react-router-dom';
 import {
   RouteParams,
-  clusterConnectConnectorEditRelativePath,
   clusterConnectConnectorRelativePath,
   clusterConnectConnectorsRelativePath,
   clusterConnectorNewRelativePath,
@@ -13,7 +12,6 @@ import useAppParams from 'lib/hooks/useAppParams';
 
 import ListPage from './List/ListPage';
 import New from './New/New';
-import Edit from './Edit/Edit';
 import DetailsPage from './Details/DetailsPage';
 
 const Connect: React.FC = () => {
@@ -23,10 +21,6 @@ const Connect: React.FC = () => {
     <Routes>
       <Route index element={<ListPage />} />
       <Route path={clusterConnectorNewRelativePath} element={<New />} />
-      <Route
-        path={clusterConnectConnectorEditRelativePath}
-        element={<Edit />}
-      />
       <Route
         path={getNonExactPath(clusterConnectConnectorRelativePath)}
         element={<DetailsPage />}
