+++ kafka-ui-react-app/src/components/Connect/Details/Config/Config.styled.ts
