+++ kafka-ui-react-app/src/components/Connect/__tests__/Connect.spec.tsx
@@ -2,12 +2,10 @@ import React from 'react';
 import { render, WithRoute } from 'lib/testHelpers';
 import { screen } from '@testing-library/react';
 import Connect from 'components/Connect/Connect';
-import { store } from 'redux/store';
 import {
   clusterConnectorsPath,
   clusterConnectorNewPath,
   clusterConnectConnectorPath,
-  clusterConnectConnectorEditPath,
   getNonExactPath,
   clusterConnectsPath,
 } from 'lib/paths';
@@ -16,7 +14,6 @@ const ConnectCompText = {
   new: 'New Page',
   list: 'List Page',
   details: 'Details Page',
-  edit: 'Edit Page',
 };
 
 jest.mock('components/Connect/New/New', () => () => (
@@ -28,9 +25,6 @@ jest.mock('components/Connect/List/ListPage', () => () => (
 jest.mock('components/Connect/Details/DetailsPage', () => () => (
   <div>{ConnectCompText.details}</div>
 ));
-jest.mock('components/Connect/Edit/Edit', () => () => (
-  <div>{ConnectCompText.edit}</div>
-));
 
 describe('Connect', () => {
   const renderComponent = (pathname: string, routePath: string) =>
@@ -38,7 +32,7 @@ describe('Connect', () => {
       <WithRoute path={getNonExactPath(routePath)}>
         <Connect />
       </WithRoute>,
-      { initialEntries: [pathname], store }
+      { initialEntries: [pathname] }
     );
 
   it('renders ListPage', () => {
@@ -63,17 +57,5 @@ describe('Connect', () => {
       clusterConnectsPath()
     );
     expect(screen.getByText(ConnectCompText.details)).toBeInTheDocument();
-  });
-
-  it('renders EditContainer', () => {
-    renderComponent(
-      clusterConnectConnectorEditPath(
-        'my-cluster',
-        'my-connect',
-        'my-connector'
-      ),
-      clusterConnectsPath()
-    );
-    expect(screen.getByText(ConnectCompText.edit)).toBeInTheDocument();
   });
 });
